final commit

+++ controllers/contactController.js
@@ -3,7 +3,7 @@ const asyncHandler = require("express-async-handler");
 const errorHandler = require('../middleware/errorHandler')
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 })
 
@@ -20,6 +20,7 @@ const createContact = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id
     });
     res.status(201).json(contact)
 })
@@ -41,26 +42,31 @@ const updateContact = asyncHandler(async (req, res) => {
         throw new Error("Contact not found");
     }
 
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
-        {new: true}
+        { new: true }
     );
     res.status(200).json(updatedContact);
 })
 
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
-  if (!contact) {
-    res.status(404);
-    throw new Error("Contact not found");
-  }
-//   if (contact.user_id.toString() !== req.user.id) {
-//     res.status(403);
-//     throw new Error("User don't have permission to update other user contacts");
-//   }
-  await Contact.deleteOne({ _id: req.params.id });
-  res.status(200).json(contact);
+    if (!contact) {
+        res.status(404);
+        throw new Error("Contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
+    }
+    await Contact.deleteOne({ _id: req.params.id });
+    res.status(200).json(contact);
 })
